requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
easyocr==1.7.2
filelock==3.20.3
fsspec==2026.1.0
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
ninja==1.13.0
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==12.1.0
pyclipper==1.4.0
PyMuPDF==1.26.7
pymupdf4llm==0.2.9
pyperclip==1.11.0
python-bidi==0.6.7
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.26.0
scipy==1.17.0
setuptools==80.9.0
shapely==2.1.2
sympy==1.14.0
tabulate==0.9.0
tifffile==2026.1.14
torch==2.9.1
torchvision==0.24.1
triton==3.5.1
typing_extensions==4.15.0
urllib3==2.6.3
